+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2019-06-18
+### Added
+- `husky` for pre-commit hook
+- `Dispatcher` for api view dispatch
+- `APIView` for inheritance
+
 ## [1.0.0] - 2019-06-07
 ### Added
 - Project inited
-- API Dispatch
 - Fetcher
 - Tests